+++ lib/screens/recipe_list_screen.dart
@@ -74,9 +74,7 @@ class _RecipeListScreenState extends ConsumerState<RecipeListScreen>
                     ),
                   );
                   if (result == true && context.mounted) {
-                    WidgetsBinding.instance.addPostFrameCallback((_) {
-                      ref.invalidate(allRecipesProvider);
-                    });
+                    ref.invalidate(allRecipesProvider);
                   }
                 },
               );
@@ -114,9 +112,7 @@ class _RecipeListScreenState extends ConsumerState<RecipeListScreen>
                 builder: (_) => const RecipeFormScreen(),
               ),
             ).then((_) {
-              WidgetsBinding.instance.addPostFrameCallback((_) {
-                ref.invalidate(allRecipesProvider);
-              });
+              if (context.mounted) ref.invalidate(allRecipesProvider);
             }),
           );
         },
@@ -164,12 +160,7 @@ class _RecipeListScreenState extends ConsumerState<RecipeListScreen>
 
     return RefreshIndicator(
       onRefresh: () async {
-        final completer = Completer<void>();
-        WidgetsBinding.instance.addPostFrameCallback((_) {
-          ref.invalidate(allRecipesProvider);
-          completer.complete();
-        });
-        await completer.future;
+        ref.invalidate(allRecipesProvider);
         await ref.read(allRecipesProvider.future);
       },
       child: ListView(
@@ -334,9 +325,7 @@ class _RecipeCard extends ConsumerWidget {
                   builder: (_) => RecipeDetailScreen(recipeId: recipe.id!),
                 ),
               ).then((_) {
-                WidgetsBinding.instance.addPostFrameCallback((_) {
-                  ref.invalidate(allRecipesProvider);
-                });
+                if (context.mounted) ref.invalidate(allRecipesProvider);
               }),
             );
           },